// src/types/ServiceUnit.type.ts
enum ServiceUnit {
  PASTOR_UNIT = 'PASTOR_UNIT',
  PROTOCOL_UNIT = 'PROTOCOL_UNIT',
  AMBIENCE_UNIT = 'AMBIENCE_UNIT',
  CELEB_KIDS_UNIT = 'CELEB_KIDS_UNIT',
  CHOIR_UNIT = 'CHOIR_UNIT',
  MEDIA_AND_TECHNICAL_UNIT = 'MEDIA_AND_TECHNICAL_UNIT',
  FOLLOW_UP_UNIT = 'FOLLOW_UP_UNIT',
  COMMS_AND_CULTURE_UNIT = 'COMMS_AND_CULTURE_UNIT',
  PHOTOGRAPHY_AND_STORYTELLING_UNIT = 'PHOTOGRAPHY_AND_STORYTELLING_UNIT',
  OUTBURST_MUSIC_GROUP_UNIT = 'OUTBURST_MUSIC_GROUP_UNIT',
  CREATIVE_UNIT = 'CREATIVE_UNIT',
  SOCIAL_MEDIA_UNIT = 'SOCIAL_MEDIA_UNIT',
  SANCTUARY_KEEPERS_UNIT = 'SANCTUARY_KEEPERS_UNIT',
}

export default ServiceUnit;